Rename running->utils, move utils->utils.utils
