+++ docs/source/api-docs/modules.rst
@@ -13,4 +13,4 @@ API documentation
    losses
    optimizers
    parts
-   running
+   utils
