+++ example_configs/speech2text/ds2_librispeech_larc_config.py
@@ -11,11 +11,12 @@ from open_seq2seq.optimizers.lr_policies import poly_decay
 base_params = {
   "random_seed": 0,
   "use_horovod": False,
-  "num_gpus": 8,
-  "batch_size_per_gpu": 16,
+  "num_gpus": 4,
+  "batch_size_per_gpu": 32,
 
   "num_epochs": 100,
 
+  "bench_start": 1000,
   "summary_frequency": 100,
   "print_loss_frequency": 10,
   "print_samples_frequency": 5000,
