+++ open_seq2seq/data/data_layer.py
@@ -5,7 +5,7 @@ import abc
 import six
 import tensorflow as tf
 import copy
-from open_seq2seq.utils import check_params
+from open_seq2seq.utils.utils import check_params
 
 
 @six.add_metaclass(abc.ABCMeta)
