+++ open_seq2seq/models/seq2seq.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2018 NVIDIA Corporation
 from __future__ import absolute_import, division, print_function
 import tensorflow as tf
-from open_seq2seq.utils import deco_print
+from open_seq2seq.utils.utils import deco_print
 from open_seq2seq.models.model import Model
